+++ editor/src/components/canvas/canvas-strategies/strategies/fragment-like-helpers.spec.browser2.tsx
