+++ editor/src/components/canvas/canvas-strategies/strategies/keyboard-set-font-size.spec.browser2.tsx
@@ -12,12 +12,12 @@ import {
   pressKey,
 } from '../../event-helpers.test-utils'
 import { EditorRenderResult, renderTestEditorWithCode } from '../../ui-jsx.test-utils'
-import { AllFragmentLikeTypes, FragmentLikeType } from './group-like-helpers'
+import { AllFragmentLikeTypes, FragmentLikeType } from './fragment-like-helpers'
 import {
   getClosingFragmentLikeTag,
   getOpeningFragmentLikeTag,
   FragmentLikeElementUid,
-} from './group-like-helpers.test-utils'
+} from './fragment-like-helpers.test-utils'
 
 describe('adjust font size with the keyboard', () => {
   describe('no font size specified', () => {
