+++ editor/src/components/common/shared-strategies/convert-to-flex-strategy.spec.browser2.tsx
@@ -7,7 +7,7 @@ import {
   selectComponentsForTest,
   setFeatureForBrowserTests,
 } from '../../../utils/utils.test-utils'
-import { getRegularNavigatorTargets } from '../../canvas/canvas-strategies/strategies/group-like-helpers.test-utils'
+import { getRegularNavigatorTargets } from '../../canvas/canvas-strategies/strategies/fragment-like-helpers.test-utils'
 import { mouseClickAtPoint, pressKey } from '../../canvas/event-helpers.test-utils'
 import {
   EditorRenderResult,
